consolidate changes into one package to resolve pertinent cyclic dependency issue
