+++ package.json
@@ -1423,7 +1423,6 @@
     "@kbn/repo-source-classifier": "link:packages/kbn-repo-source-classifier",
     "@kbn/repo-source-classifier-cli": "link:packages/kbn-repo-source-classifier-cli",
     "@kbn/security-api-integration-helpers": "link:x-pack/test/security_api_integration/packages/helpers",
-    "@kbn/security-role-management-model-mocks": "link:x-pack/packages/security/role_management_model_mocks",
     "@kbn/serverless-storybook-config": "link:packages/serverless/storybook/config",
     "@kbn/some-dev-log": "link:packages/kbn-some-dev-log",
     "@kbn/sort-package-json": "link:packages/kbn-sort-package-json",
