+++ x-pack/packages/security/role_management_model/src/kibana_privileges.test.ts
@@ -8,10 +8,7 @@
 import { KibanaPrivilege } from './kibana_privilege';
 import { KibanaPrivileges, isGlobalPrivilegeDefinition } from './kibana_privileges';
 import type { RoleKibanaPrivilege } from '@kbn/security-plugin-types-common';
-import {
-  createRawKibanaPrivileges,
-  kibanaFeatures,
-} from '@kbn/security-role-management-model-mocks';
+import { createRawKibanaPrivileges, kibanaFeatures } from './__fixtures__';
 
 describe('kibana_privilege', () => {
   describe('isGlobalPrivilegeDefinition', () => {
