+++ x-pack/packages/security/role_management_model/tsconfig.json
@@ -10,6 +10,7 @@
     "@kbn/features-plugin",
     "@kbn/security-plugin-types-common",
     "@kbn/security-authorization-core",
-    "@kbn/security-role-management-model-mocks"
+    "@kbn/security-role-management-model",
+    "@kbn/licensing-plugin",
   ]
 }
