--- x-pack/packages/security/role_management_model_mocks/README.md
@@ -1,3 +1,0 @@
-# @kbn/security-role-management-mocks
-
-Contains business logic for RBAC administration within Kibana.
