--- x-pack/packages/security/role_management_model_mocks/kibana.jsonc
@@ -1,5 +1,0 @@
-{
-  "type": "test-helper",
-  "id": "@kbn/security-role-management-model-mocks",
-  "owner": "@elastic/kibana-security"
-}
