--- x-pack/packages/security/role_management_model_mocks/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "@kbn/security-role-management-model-mocks",
-  "private": true,
-  "version": "1.0.0",
-  "license": "Elastic License 2.0"
-}
