--- x-pack/packages/security/role_management_model_mocks/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "target/types",
-    "types": ["jest", "node", "react"]
-  },
-  "include": ["**/*.ts", "**/*.tsx"],
-  "exclude": ["target/**/*"],
-  "kbn_references": [
-    "@kbn/features-plugin",
-    "@kbn/security-plugin-types-common",
-    "@kbn/licensing-plugin",
-    "@kbn/security-role-management-model",
-    "@kbn/security-authorization-core"
-  ]
-}
