+++ x-pack/plugins/security/public/management/roles/edit_role/edit_role_page.test.tsx
@@ -19,7 +19,7 @@ import { dataViewPluginMocks } from '@kbn/data-views-plugin/public/mocks';
 import { KibanaFeature } from '@kbn/features-plugin/public';
 import { KibanaContextProvider } from '@kbn/kibana-react-plugin/public';
 import { REMOTE_CLUSTERS_PATH } from '@kbn/remote-clusters-plugin/public';
-import { createRawKibanaPrivileges } from '@kbn/security-role-management-model-mocks';
+import { createRawKibanaPrivileges } from '@kbn/security-role-management-model/src/__fixtures__';
 import type { Space } from '@kbn/spaces-plugin/public';
 import { spacesManagerMock } from '@kbn/spaces-plugin/public/spaces_manager/mocks';
 import { getUiApi } from '@kbn/spaces-plugin/public/ui_api';
