+++ x-pack/plugins/security/public/management/roles/edit_role/privileges/kibana/feature_table/feature_table.test.tsx
@@ -13,7 +13,7 @@ import {
   createFeature,
   createKibanaPrivileges,
   kibanaFeatures,
-} from '@kbn/security-role-management-model-mocks';
+} from '@kbn/security-role-management-model/src/__fixtures__';
 import { findTestSubject, mountWithIntl } from '@kbn/test-jest-helpers';
 
 import { getDisplayedFeaturePrivileges } from './__fixtures__';
