+++ x-pack/plugins/security/public/management/roles/edit_role/privileges/kibana/privilege_form_calculator/privilege_form_calculator.test.ts
@@ -5,7 +5,10 @@
  * 2.0.
  */
 
-import { createKibanaPrivileges, kibanaFeatures } from '@kbn/security-role-management-model-mocks';
+import {
+  createKibanaPrivileges,
+  kibanaFeatures,
+} from '@kbn/security-role-management-model/src/__fixtures__';
 
 import { PrivilegeFormCalculator } from './privilege_form_calculator';
 import type { Role } from '../../../../../../../common';
