+++ x-pack/plugins/security/public/management/roles/edit_role/privileges/kibana/space_aware_privilege_section/privilege_space_form.test.tsx
@@ -12,7 +12,7 @@ import {
   createFeature,
   createKibanaPrivileges,
   kibanaFeatures,
-} from '@kbn/security-role-management-model-mocks';
+} from '@kbn/security-role-management-model/src/__fixtures__';
 import type { Space } from '@kbn/spaces-plugin/public';
 import { findTestSubject, mountWithIntl } from '@kbn/test-jest-helpers';
