+++ x-pack/plugins/security/public/management/roles/edit_role/privileges/kibana/space_aware_privilege_section/privilege_space_table.test.tsx
@@ -11,7 +11,7 @@ import React from 'react';
 
 import { KibanaFeature } from '@kbn/features-plugin/public';
 import type { Role, RoleKibanaPrivilege } from '@kbn/security-plugin-types-common';
-import { createKibanaPrivileges } from '@kbn/security-role-management-model-mocks';
+import { createKibanaPrivileges } from '@kbn/security-role-management-model/src/__fixtures__';
 import { findTestSubject, mountWithIntl } from '@kbn/test-jest-helpers';
 
 import { PrivilegeDisplay } from './privilege_display';
